Update to League Management View
includes updates to the form entry to be contained with in a container to give nice borders.
