+++ src/DartLeague/DartLeague.Web/Areas/Manage/Controllers/HomeController.cs
@@ -22,32 +22,9 @@ namespace DartLeague.Web.Areas.Manage.Controllers
             _browsableFileService = browsableFileService;
         }
 
-        public async Task<IActionResult> Index()
+        public IActionResult Index()
         {
-            var links = await GetLeagueLinks();
-            return View(new HomeManagementViewModel
-            {
-                LeagueLinks = links
-            });
-        }
-
-        private async Task<List<LeagueLinksListViewModel>> GetLeagueLinks()
-        {
-            var list = new List<LeagueLinksListViewModel>();
-            foreach (var link in await _leagueContext.LeagueLinks.OrderBy(x => x.Order).ToListAsync())
-            {
-                var linkView = new LeagueLinksListViewModel
-                {
-                    Id = link.Id,
-                    Title = link.Title,
-                    LinkType = link.LinkType == 1 ? "Url" : "File",
-                    Url = link.Url,
-                    Order = link.Order
-                };
-                list.Add(linkView);
-            }
-
-            return list;
+            return RedirectToAction("Index", "LeagueLink");
         }
     }
 }
