+++ src/DartLeague/DartLeague.Web/Areas/Manage/Models/SeasonListViewModel.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace DartLeague.Web.Areas.Manage.Models
+{
+    public class SeasonListViewModel
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public bool Active { get; set; }
+
+    }
+}
