+++ src/DartLeague/DartLeague.Web/Areas/Manage/Views/Member/Create.cshtml
@@ -3,16 +3,17 @@
     ViewData["Title"] = "Add Member";
 
 }
-
-<h2>Add Member</h2>
-<form asp-controller="Member" asp-action="Create" method="post">
-    <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-    @Html.Partial("_memberFields")
-    <div>
-        <button type="submit" class="btn btn-default"><i class="fa fa-floppy-o fa-lg"></i> Add</button>
-        <a class="btn btn-danger" asp-controller="Member" asp-action="Index">Cancel</a>
-    </div>
-</form>
+<div class="container">
+    <h2>Add Member</h2>
+    <form asp-controller="Member" asp-action="Create" method="post">
+        <div asp-validation-summary="ModelOnly" class="text-danger"></div>
+        @Html.Partial("_memberFields")
+        <div>
+            <button type="submit" class="btn btn-default"><i class="fa fa-floppy-o fa-lg"></i> Add</button>
+            <a class="btn btn-danger" asp-controller="Member" asp-action="Index">Cancel</a>
+        </div>
+    </form>
+</div>
 
 @section scripts{
         <script>
