+++ src/DartLeague/DartLeague.Web/Areas/Manage/Views/Member/Index.cshtml
@@ -5,41 +5,47 @@
 @using IdentityServer4.Extensions
 @model DartLeague.Web.Areas.Manage.Models.MembersListViewModel
 
-<div class="single-form">
-    <div class="panel panel-default">
-        <div class="panel-heading"><a class="btn btn-success btn-sm" asp-area="Manage" asp-controller="Member" asp-action="Create" title="Add Member"><i class="fa fa-plus-square"></i></a> <strong>Members</strong></div>
-        <table class="table table-condensed table-striped table-hover">
-            <thead>
-                <tr>
-                    <th class="hidden-sm hidden-xs" width="5px"></th>
-                    <th>Name</th>
-                    <th>Email</th>
-                    <th>Cell Phone</th>
-                    <th class="hidden-xs">Shirt</th>
-                    <th class="hidden-sm hidden-xs" width="5px"></th>
-                </tr>
-            </thead>
-            <tbody>
-                @if (!Model.Members.Any())
-                {
-                    <tr class="warning"><td colspan="6">Currently No Members</td></tr>
-                }
-                else
-                {
-                    foreach (var member in Model.Members)
-                    {
+@await Html.PartialAsync("_ManagementHeader")
+
+<div class="container">
+    <div class="row">
+        <div class="col-lg-12">
+            <div class="panel panel-default">
+                <div class="panel-heading"><a class="btn btn-success btn-sm" asp-area="Manage" asp-controller="Member" asp-action="Create" title="Add Member"><i class="fa fa-plus-square"></i></a> <strong>Members</strong></div>
+                <table class="table table-condensed table-striped table-hover">
+                    <thead>
                         <tr>
-                            <td class="hidden-sm hidden-xs"><a class="btn btn-link" asp-area="Manage" asp-controller="Member" asp-action="Edit" asp-route-id="@member.Id"><i class="fa fa-pencil-square-o fa-lg"></i></a></td>
-                            <td><a class="btn btn-link" asp-area="Manage" asp-controller="Member" asp-action="Edit" asp-route-id="@member.Id">@($"{member.FirstName} {member.LastName}")</a></td>
-                            <td>@(member.Email.IsNullOrEmpty() ? "none" : member.Email)</td>
-                            <td>@member.CellPhone</td>
-                            <td class="hidden-xs">@member.ShirtSize</td>
-                            <td class="hidden-sm hidden-xs"><a class="btn btn-link confirmation" asp-area="Manage" asp-controller="Member" asp-action="Delete" asp-route-id="@member.Id" title="Delete"><i class="fa fa-trash-o fa-lg"></i></a></td>
+                            <th class="hidden-sm hidden-xs" width="5px"></th>
+                            <th>Name</th>
+                            <th>Email</th>
+                            <th>Cell Phone</th>
+                            <th class="hidden-xs">Shirt</th>
+                            <th class="hidden-sm hidden-xs" width="5px"></th>
                         </tr>
-                    }
-                }
-            </tbody>
-        </table>
+                    </thead>
+                    <tbody>
+                        @if (!Model.Members.Any())
+                        {
+                            <tr class="warning"><td colspan="6">Currently No Members</td></tr>
+                        }
+                        else
+                        {
+                            foreach (var member in Model.Members)
+                            {
+                                <tr>
+                                    <td class="hidden-sm hidden-xs"><a class="btn btn-link" asp-area="Manage" asp-controller="Member" asp-action="Edit" asp-route-id="@member.Id"><i class="fa fa-pencil-square-o fa-lg"></i></a></td>
+                                    <td><a class="btn btn-link" asp-area="Manage" asp-controller="Member" asp-action="Edit" asp-route-id="@member.Id">@($"{member.FirstName} {member.LastName}")</a></td>
+                                    <td>@(member.Email.IsNullOrEmpty() ? "none" : member.Email)</td>
+                                    <td>@member.CellPhone</td>
+                                    <td class="hidden-xs">@member.ShirtSize</td>
+                                    <td class="hidden-sm hidden-xs"><a class="btn btn-link confirmation" asp-area="Manage" asp-controller="Member" asp-action="Delete" asp-route-id="@member.Id" title="Delete"><i class="fa fa-trash-o fa-lg"></i></a></td>
+                                </tr>
+                            }
+                        }
+                    </tbody>
+                </table>
+            </div>
+        </div>
     </div>
 </div>
 @section scripts{
