+++ src/DartLeague/DartLeague.Web/Areas/Manage/Views/Sponsor/Edit.cshtml
@@ -1,15 +1,17 @@
 ﻿@{
     ViewData["Title"] = "Edit Sponsor";
 }
-<h2>Edit Sponsor</h2>
-<form asp-controller="Sponsor" asp-action="Edit" method="post">
-    <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-    @Html.Partial("_sponsorFields")
-    <div>
-        <button type="submit" class="btn btn-default"><i class="fa fa-floppy-o fa-lg"></i> Save</button>
-        <a class="btn btn-danger" asp-area="Manage" asp-controller="Sponsor" asp-action="Index">Cancel</a>
-    </div>
-</form>
+<div class="container">
+    <h2>Edit Sponsor</h2>
+    <form asp-controller="Sponsor" asp-action="Edit" method="post">
+        <div asp-validation-summary="ModelOnly" class="text-danger"></div>
+        @Html.Partial("_sponsorFields")
+        <div>
+            <button type="submit" class="btn btn-default"><i class="fa fa-floppy-o fa-lg"></i> Save</button>
+            <a class="btn btn-danger" asp-area="Manage" asp-controller="Sponsor" asp-action="Index">Cancel</a>
+        </div>
+    </form>
+</div>
 @section scripts{
     <script>
         (function(d){
