+++ src/DartLeague/DartLeague.Web/Areas/Manage/Views/Sponsor/Index.cshtml
@@ -4,44 +4,54 @@
 
 @model DartLeague.Web.Areas.Manage.Models.SponsorListViewModel
 
-<div class="single-form">
-    <div class="panel panel-default">
-        <div class="panel-heading"><a class="btn btn-success btn-sm" asp-area="Manage" asp-controller="Sponsor" asp-action="Create" title="Add Sponsor">
-            <i class="fa fa-plus-square"></i>
-        </a> <strong>Sponsors</strong></div>
-        <div class="table-responsive">
-            <table class="table table-striped table-hover">
-                <thead>
-                <tr>
-                    <th width="5px"></th>
-                    <th>Name</th>
-                    <th>Type</th>
-                    <th>Contact</th>
-                    <th width="5px"></th>
-                </tr>
-                </thead>
-                <tbody>
-                @if (!Model.Sponsors.Any())
-                {
-                    <tr class="warning"><td colspan="5">Currently No Sponsors</td></tr>
-                }
-                else
-                {
-                    foreach (var sponsor in Model.Sponsors)
-                    {
-                        <tr>
-                            <td><a class="btn btn-link" asp-area="Manage" asp-controller="Sponsor" asp-action="Edit" asp-route-id="@sponsor.Id" title="Edit @sponsor.Name">
-                                <i class="fa fa-pencil-square-o fa-lg"></i>
-                            </a></td>
-                            <td>@sponsor.Name</td>
-                            <td>@sponsor.Type</td>
-                            <td>@sponsor.ContactName</td>
-                            <td><a class="btn btn-link confirmation" asp-area="Manage" asp-controller="Sponsor" asp-action="Delete" asp-route-id="@sponsor.Id" title="Delete"><i class="fa fa-trash-o fa-lg"></i></a>
-                        </tr>
-                    }
-                }
-                </tbody>
-            </table>
+@await  Html.PartialAsync("_ManagementHeader")
+
+<div class="container">
+    <div class="row">
+        <div class="col-lg-12">
+            <div class="panel panel-default">
+                <div class="panel-heading">
+                    <a class="btn btn-success btn-sm" asp-area="Manage" asp-controller="Sponsor" asp-action="Create" title="Add Sponsor">
+                        <i class="fa fa-plus-square"></i>
+                    </a> <strong>Sponsors</strong>
+                </div>
+                <div class="table-responsive">
+                    <table class="table table-striped table-hover">
+                        <thead>
+                            <tr>
+                                <th width="5px"></th>
+                                <th>Name</th>
+                                <th>Type</th>
+                                <th>Contact</th>
+                                <th width="5px"></th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            @if (!Model.Sponsors.Any())
+                            {
+                                <tr class="warning"><td colspan="5">Currently No Sponsors</td></tr>
+                            }
+                            else
+                            {
+                                foreach (var sponsor in Model.Sponsors)
+                                {
+                                    <tr>
+                                        <td>
+                                            <a class="btn btn-link" asp-area="Manage" asp-controller="Sponsor" asp-action="Edit" asp-route-id="@sponsor.Id" title="Edit @sponsor.Name">
+                                                <i class="fa fa-pencil-square-o fa-lg"></i>
+                                            </a>
+                                        </td>
+                                        <td>@sponsor.Name</td>
+                                        <td>@sponsor.Type</td>
+                                        <td>@sponsor.ContactName</td>
+                                        <td><a class="btn btn-link confirmation" asp-area="Manage" asp-controller="Sponsor" asp-action="Delete" asp-route-id="@sponsor.Id" title="Delete"><i class="fa fa-trash-o fa-lg"></i></a>
+                                    </tr>
+                                }
+                            }
+                        </tbody>
+                    </table>
+                </div>
+            </div>
         </div>
     </div>
 </div>
