+++ src/DartLeague/DartLeague.Web/Areas/Site/Views/DartEvent/Create.cshtml
@@ -4,17 +4,17 @@
     ViewBag.Title = "Add Event";
     Layout = "_Layout";
 }
-
-<h2>Create Dart Event</h2>
-<form asp-controller="DartEvent" enctype="multipart/form-data" asp-action="Create" method="post">
-    @Html.Partial("_dartEventFields")
-    <div>
-        <button type="submit" class="btn btn-default"><i class="fa fa-floppy-o fa-lg"></i> Add</button>
-        <a class="btn btn-danger" asp-controller="DartEvent" asp-action="Index">Cancel</a>
-    </div>
-</form>
-
-@section Scripts{
+<div class="container">
+    <h2>Create Dart Event</h2>
+    <form asp-controller="DartEvent" enctype="multipart/form-data" asp-action="Create" method="post">
+        @Html.Partial("_dartEventFields")
+        <div>
+            <button type="submit" class="btn btn-default"><i class="fa fa-floppy-o fa-lg"></i> Add</button>
+            <a class="btn btn-danger" asp-controller="DartEvent" asp-action="Index">Cancel</a>
+        </div>
+    </form>
+</div>
+@section scripts{
     <script>
         (function(d) {
             d.getElementsByName('Name')[0].focus();
